span.SetName() method added
