+++ span_implementation.go
@@ -19,6 +19,12 @@ func (s *spanImpl) Context() model.SpanContext {
 	return s.SpanContext
 }
 
+func (s *spanImpl) SetName(name string) {
+	s.mtx.Lock()
+	s.Name = name
+	s.mtx.Unlock()
+}
+
 func (s *spanImpl) SetRemoteEndpoint(e *model.Endpoint) {
 	s.mtx.Lock()
 	if e == nil {
@@ -30,7 +36,6 @@ func (s *spanImpl) SetRemoteEndpoint(e *model.Endpoint) {
 	s.mtx.Unlock()
 }
 
-// Annotate adds a new Annotation to the Span.
 func (s *spanImpl) Annotate(t time.Time, value string) {
 	a := model.Annotation{
 		Timestamp: t,
@@ -42,9 +47,6 @@ func (s *spanImpl) Annotate(t time.Time, value string) {
 	s.mtx.Unlock()
 }
 
-// Tag sets Tag with given key and value to the Span. If key already exists in
-// the span the value will be overridden except for error tags where the first
-// value is persisted.
 func (s *spanImpl) Tag(key, value string) {
 	s.mtx.Lock()
 
@@ -59,7 +61,6 @@ func (s *spanImpl) Tag(key, value string) {
 	s.mtx.Unlock()
 }
 
-// Finish the span and send to reporter.
 func (s *spanImpl) Finish() {
 	if atomic.CompareAndSwapInt32(&s.mustCollect, 1, 0) {
 		s.Duration = time.Since(s.Timestamp)
